refactor: wrap metrics listener into fastify plugin
